probably add firewalld configuration for quasselcore
This installs a firewalld service file named quasselcore.xml that opens
TCP port 4242, and opens up that quasselcore service on the default
zone.

I set up the actual firewall configuration manually, so I'm not sure
the commands here actually work. But it transfers the file and runs the
module without exploding, so I'll assume it's OK for now.
(Fortunately the failure mode is pretty obvious...the Quassel client
will say "Network unreachable.")